+++ src/runtime/internal/types.ts
@@ -7,9 +7,12 @@ export type RuntimeCookieOptions = {
 
 export type RuntimeMessages = Record<string, Record<string, unknown>>
 
+export type RuntimeOnMissing = (path: string, locale: string, fallbackLocale: string) => string | undefined
+
 export type RuntimeOptions = {
     cookie: RuntimeCookieOptions
     defaultLocale: string
+    detectLocale: boolean
     dir: string
     fallbackLocale: string
     locales: Array<{
@@ -17,5 +20,6 @@ export type RuntimeOptions = {
         file: `${string}.ts`
         name: string
     }>
+    onMissing?: RuntimeOnMissing
     warnOnMissing: boolean
 }