+++ src/types.ts
@@ -11,11 +11,15 @@ export type NuxtI18nLocale = {
     name: string
 }
 
+export type NuxtI18nOnMissing = (path: string, locale: string, fallbackLocale: string) => string | undefined
+
 export interface NuxtI18nOptions {
     cookie?: NuxtI18nCookieOptions
     defaultLocale?: string
+    detectLocale?: boolean
     dir?: string
     fallbackLocale?: string
     locales?: NuxtI18nLocale[]
+    onMissing?: NuxtI18nOnMissing
     warnOnMissing?: boolean
 }