+++ test/fixtures/basic/nuxt.config.ts
@@ -8,6 +8,7 @@ export default defineNuxtConfig({
             name: 'preferred-locale',
         },
         defaultLocale: 'es',
+        detectLocale: true,
         dir: 'i18n',
         fallbackLocale: 'es',
         locales: [
@@ -22,6 +23,7 @@ export default defineNuxtConfig({
                 name: 'English',
             },
         ],
+        onMissing: (path, locale) => `[missing:${locale}] ${path}`,
         warnOnMissing: true,
     },
     modules: [MyModule],